+++ src/components/FigmaLink/index.tsx
@@ -2,7 +2,7 @@ type Props = {
     url: string;
 }
 
-export default function FigmaLink({url}: Props) {
+export const FigmaLink = ({url}: Props) => {
   return (
     <a href={url} className="absolute bottom-0 right-0" target={"_blank"} title="Abrir no Figma">
         <img src="/images/figmalogo.png" alt="Figma" className="h-10 w-10" />
