+++ src/components/SectionTitle/index.tsx
@@ -3,7 +3,7 @@ type Props = {
   className?: string;
 };
 
-export default function SectionTitle({ title, className }: Props) {
+export const SectionTitle = ({ title, className }: Props) => {
   return (
     <span className={`text-white tracking-[0.3em] font-medium ${className}`}>
       {title.toUpperCase()}
