+++ src/components/ViewMore/index.tsx
@@ -6,7 +6,7 @@ type Props = {
   className?: string;
 }
 
-export default function ViewMore({url, blank = false, className}: Props) {
+export const ViewMore = ({url, blank = false, className}: Props) => {
   return (
     <a href={url} target={blank ? '_blank' : '_self'} className={`flex flex-row items-center ${className}`}>Ver Mais <CaretRight /></a>
   )
