+++ src/components/index.ts
@@ -6,3 +6,15 @@ export * from './LanguageSelector';
 export * from './BackgroundElement';
 export * from './Button';
 export * from './ProjectCard';
+export * from './Challenges';
+export * from './SectionTitle';
+export * from './SkillDescription';
+export * from './FigmaLink';
+export * from './Footer';
+export * from './ProjectListFull';
+export * from './ProjectsList';
+export * from './ViewMore';
+export * from './SkillBlock';
+export * from './SkillLevel';
+export * from './SmallHeader';
+export * from './SubmitButton';
