+++ src/pages/Home/index.tsx
@@ -1,5 +1,4 @@
-import { Header } from '../../components';
-import Footer from "../../components/Footer/Footer";
+import { Header, Footer } from '../../components';
 import { Wellcome, Challenges, Contact, Projects, AboutMe, Skills } from '../../sections/';
 
 export const Home = () => {
